requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
natsort==7.0.1
packaging==20.3
pep517==0.8.2
progress==1.5
PyJWT==1.7.1
pyparsing==2.4.6
python-dotenv==0.14.0
python-firebase==1.2
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
sty==1.0.0b12
termcolor==1.1.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
